페이징함수 객체화 (at CommonService)
