+++ HONBAB/src/main/java/com/web/honbab/common/service/CommonService.java
@@ -1,0 +1,9 @@
+package com.web.honbab.common.service;
+
+import org.springframework.ui.Model;
+
+public interface CommonService {
+	
+	public int[] paging(Model model, int num, int allCount);
+
+}
